requirements.txt
aniso8601==3.0.2
astroid==2.1.0
colorama==0.4.3
Django==2.1.3
django-graphql-jwt==0.3.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
promise==2.2.1
PyJWT==1.7.1
pylint==2.2.2
pytz==2018.7
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
